src/requirements.txt
Django==3.1.4
django-cors-headers==3.5.0
django-environ==0.4.5
djangorestframework==3.12.2
ipython==7.19.0
psycopg2-binary==2.8.6
requests==2.25.0
